implemented last method + tests
