+++ Assignment3.Tests/WizardTests.cs
@@ -29,16 +29,7 @@ namespace BDSA2021.Assignment03.Tests
             );
         }
 
-        [Fact]
-        public void getWizardsFromAuthor_Given_Rowling_Returns_HarryPotter_Voldemort_Dobby_HermioneGranger_RonWeasley_GinnyWeasley_FredWeasley_GeorgeWeasley()
-        {
-            var wizards = Wizard.Wizards.Value;
-
-            string[] output = Queries.getWizardsFromAuthor("Rowling",wizards).ToArray();
-
-            Assert.Equal(output,new string[]{"Harry Potter", "Voldemort", "Dobby", "Hermione Granger", "Ron Weasley","Ginny Weasley","Fred Weasley","George Weasley"});
         
-        }
         [Fact]
         public void Extention_Given_Rowling_Returns_HarryPotter_Voldemort_Dobby_HermioneGranger_RonWeasley_GinnyWeasley_FredWeasley_GeorgeWeasley()
         {
@@ -46,16 +37,7 @@ namespace BDSA2021.Assignment03.Tests
             var output = w.Where(w=>w.Creator.Contains("Rowling")).Select(w=>w.Name);
             Assert.Equal(output,new string[]{"Harry Potter", "Voldemort", "Dobby", "Hermione Granger", "Ron Weasley","Ginny Weasley","Fred Weasley","George Weasley"});
         }
-
-
-        [Fact]
-        public void getYearOfFirstSithLord_given_al_wizards_returns_1900()
-        {
-            var wizards = Wizard.Wizards.Value;
-            int output = Queries.getYearOfFirstSithLord(wizards);
-
-            Assert.Equal(output, 1900);
-        }
+        
 
         public void Extention_given_al_wizards_returns_1900()
         {
@@ -66,26 +48,7 @@ namespace BDSA2021.Assignment03.Tests
         }
 
 
-        [Fact]
-        public void getHarryPotterWizards_returns_allHarryPotterWizardsAsTuples()
-        {
-            var wizards = Wizard.Wizards.Value;
-            var output = Queries.getHarryPotterWizards(wizards);
-            
-            var expected = new List<(string, int)>
-            {
-                ("Harry Potter", 1997),
-                ("Voldemort", 1997),
-                ("Dobby", 1997),
-                ("Hermione Granger", 1997),
-                ("Ron Weasley", 1997),
-                ("Ginny Weasley", 1997),
-                ("Fred Weasley", 1997),
-                ("George Weasley", 1997)
-            };
-
-            Assert.Equal(output, expected);
-        }
+       
         [Fact]
         public void Extention_returns_allHarryPotterWizardsAsTuples()
         {
@@ -109,24 +72,17 @@ namespace BDSA2021.Assignment03.Tests
 
         
 
-        [Fact]
-        public void getWizardNames_returns_AllWizardsNames_reverseOrder_groupedByCreator()
-        {
-            var wizards = Wizard.Wizards.Value;
-            var output = Queries.getWizardNames(wizards);
-
-            Assert.Equal(output, new string[]{"Merlin","Sauron","Gandalf","Voldemort","Ron Weasley","Hermione Granger","Harry Potter","Ginny Weasley","George Weasley","Fred Weasley","Dobby","Melly Sandra","Darth Vader","Darth Maul"}); 
-        }
+       
 
         [Fact]
         public void Extention_returns_AllWizardsNames_reverseOrder_groupedByCreator()
         {
             //from w in wizards group w by new {w.Creator, w.Name} into g orderby g.Key.Creator descending, g.Key.Name descending select g.Key.Name;
            
-           /*  var w = Wizard.Wizards.Value;
-            var output = w.GroupBy(w=> new {w.Creator, w.Name},(key, group) => new {key1 = key.Creator, key2= key.Name}).OrderByDescending(w=>w.Creator).ThenByDescending(w=>w.Name).Select(w=>w.Name);
+            var w = Wizard.Wizards.Value;
+            var output = w.GroupBy(w=> new {w.Creator, w.Name}).OrderByDescending(w=>w.Key.Creator).ThenByDescending(w=>w.Key.Name).Select(w=>w.Key.Name);
 
-            Assert.Equal(output, new string[]{"Merlin","Sauron","Gandalf","Voldemort","Ron Weasley","Hermione Granger","Harry Potter","Ginny Weasley","George Weasley","Fred Weasley","Dobby","Melly Sandra","Darth Vader","Darth Maul"});  */
+            Assert.Equal(output, new string[]{"Merlin","Sauron","Gandalf","Voldemort","Ron Weasley","Hermione Granger","Harry Potter","Ginny Weasley","George Weasley","Fred Weasley","Dobby","Melly Sandra","Darth Vader","Darth Maul"});  
         }
